style: change proto type

+++ src/Application/DTO/GetRatingDto.cs
@@ -4,7 +4,7 @@ public class GetRatingDto
 {
     public long SubjectId { get; init; }
 
-    public decimal Average { get; init; }
+    public double Average { get; init; }
 
     public long Count { get; init; }
 }
